Change `Entity` and other generational IDs to use `u64` instead of `u32`

+++ crates/vizia_core/src/entity.rs
@@ -16,6 +16,6 @@ use vizia_id::{
 /// [`Restyle`]: crate::prelude::WindowEvent::Restyle
 /// [`Redraw`]: crate::prelude::WindowEvent::Redraw
 #[derive(Clone, Copy, PartialEq, Eq, Hash)]
-pub struct Entity(u32);
+pub struct Entity(u64);
 
 impl_generational_id!(Entity);
